requirements.txt
Django==1.9
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
Pillow==3.4.1
